feat: use sqlite database for file uploads; split up server into different route handlers
